requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
confluent-kafka==2.8.2
dnspython==2.7.0
Faker==36.2.2
h2==2.6.2
hpack==3.0.0
hyper==0.7.0
hyperframe==3.2.0
idna==3.10
kafka-python==2.0.6
ksql==0.10.2
pyflink==1.0
pymongo==4.11.2
requests==2.32.3
six==1.17.0
tzdata==2025.1
urllib3==2.3.0
